# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PropertyType = typeof PropertyType[keyof typeof PropertyType];


export const PropertyType = {
  apartamento: 'apartamento',
  casa: 'casa',
  terreno: 'terreno',
  comercial: 'comercial',
  cobertura: 'cobertura',
  rural: 'rural',
} as const;

export type PropertyStatus = typeof PropertyStatus[keyof typeof PropertyStatus];


export const PropertyStatus = {
  captado: 'captado',
  em_negociacao: 'em_negociacao',
  disponivel: 'disponivel',
  alugado: 'alugado',
} as const;

/**
 * Fixed capture zone within Porto Alegre
 */
export type Zona = typeof Zona[keyof typeof Zona];


export const Zona = {
  centro_menino_deus: 'centro_menino_deus',
  moinhos_floresta: 'moinhos_floresta',
} as const;

/**
 * Fixed list of neighborhoods allowed for capture, grouped by Zona
 */
export type Bairro = typeof Bairro[keyof typeof Bairro];


export const Bairro = {
  centro: 'centro',
  cidade_baixa: 'cidade_baixa',
  farroupilha: 'farroupilha',
  bom_fim: 'bom_fim',
  independencia: 'independencia',
  santana: 'santana',
  praia_de_belas: 'praia_de_belas',
  menino_deus: 'menino_deus',
  azenha: 'azenha',
  santa_tereza: 'santa_tereza',
  medianeira: 'medianeira',
  santo_antonio: 'santo_antonio',
  floresta: 'floresta',
  sao_geraldo: 'sao_geraldo',
  higienopolis: 'higienopolis',
  sao_joao: 'sao_joao',
  santa_maria_goretti: 'santa_maria_goretti',
  jardim_floresta: 'jardim_floresta',
  jardim_sao_pedro: 'jardim_sao_pedro',
  humaita: 'humaita',
  navegantes: 'navegantes',
  anchieta: 'anchieta',
  farrapos: 'farrapos',
  moinhos_de_vento: 'moinhos_de_vento',
  auxiliadora: 'auxiliadora',
  montserrat: 'montserrat',
  rio_branco: 'rio_branco',
  bela_vista: 'bela_vista',
  santa_cecilia: 'santa_cecilia',
} as const;

export interface Property {
  id: number;
  title: string;
  zona: Zona;
  neighborhood: Bairro;
  address: string;
  /** Monthly rent value in BRL */
  price: number;
  propertyType: PropertyType;
  status: PropertyStatus;
  bedrooms: number;
  bathrooms: number;
  areaM2: number;
  ownerName: string;
  ownerPhone: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PropertyInput {
  /** @minLength 1 */
  title: string;
  zona: Zona;
  neighborhood: Bairro;
  /** @minLength 1 */
  address: string;
  /** Monthly rent value in BRL */
  price: number;
  propertyType: PropertyType;
  status: PropertyStatus;
  bedrooms: number;
  bathrooms: number;
  areaM2: number;
  /** @minLength 1 */
  ownerName: string;
  /** @minLength 1 */
  ownerPhone: string;
  notes?: string;
}

export interface PropertyUpdate {
  /** @minLength 1 */
  title?: string;
  zona?: Zona;
  neighborhood?: Bairro;
  /** @minLength 1 */
  address?: string;
  /** Monthly rent value in BRL */
  price?: number;
  propertyType?: PropertyType;
  status?: PropertyStatus;
  bedrooms?: number;
  bathrooms?: number;
  areaM2?: number;
  /** @minLength 1 */
  ownerName?: string;
  /** @minLength 1 */
  ownerPhone?: string;
  notes?: string;
}

export interface DashboardSummary {
  totalProperties: number;
  capturedThisMonth: number;
  /** Count of properties currently available for rent */
  availableForRent: number;
  /** Average monthly rent value across all captured properties */
  averageRent: number;
}

export interface PropertyTypeCount {
  propertyType: PropertyType;
  count: number;
}

export interface ZonaCount {
  zona: Zona;
  count: number;
}

export type ListPropertiesParams = {
zona?: Zona;
neighborhood?: Bairro;
maxPrice?: number;
propertyType?: PropertyType;
};

export type GetRecentPropertiesParams = {
limit?: number;
};