# lib/api-zod/src/generated/types/bairro.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Fixed list of neighborhoods allowed for capture, grouped by Zona
 */
export type Bairro = typeof Bairro[keyof typeof Bairro];


export const Bairro = {
  centro: 'centro',
  cidade_baixa: 'cidade_baixa',
  farroupilha: 'farroupilha',
  bom_fim: 'bom_fim',
  independencia: 'independencia',
  santana: 'santana',
  praia_de_belas: 'praia_de_belas',
  menino_deus: 'menino_deus',
  azenha: 'azenha',
  santa_tereza: 'santa_tereza',
  medianeira: 'medianeira',
  santo_antonio: 'santo_antonio',
  floresta: 'floresta',
  sao_geraldo: 'sao_geraldo',
  higienopolis: 'higienopolis',
  sao_joao: 'sao_joao',
  santa_maria_goretti: 'santa_maria_goretti',
  jardim_floresta: 'jardim_floresta',
  jardim_sao_pedro: 'jardim_sao_pedro',
  humaita: 'humaita',
  navegantes: 'navegantes',
  anchieta: 'anchieta',
  farrapos: 'farrapos',
  moinhos_de_vento: 'moinhos_de_vento',
  auxiliadora: 'auxiliadora',
  montserrat: 'montserrat',
  rio_branco: 'rio_branco',
  bela_vista: 'bela_vista',
  santa_cecilia: 'santa_cecilia',
} as const;