# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalProperties: number;
  capturedThisMonth: number;
  /** Count of properties currently available for rent */
  availableForRent: number;
  /** Average monthly rent value across all captured properties */
  averageRent: number;
}