# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Bairro } from './bairro';
import type { PropertyStatus } from './propertyStatus';
import type { PropertyType } from './propertyType';
import type { Zona } from './zona';

export interface Property {
  id: number;
  title: string;
  zona: Zona;
  neighborhood: Bairro;
  address: string;
  /** Monthly rent value in BRL */
  price: number;
  propertyType: PropertyType;
  status: PropertyStatus;
  bedrooms: number;
  bathrooms: number;
  areaM2: number;
  ownerName: string;
  ownerPhone: string;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}