# lib/api-zod/src/generated/types/propertyInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Bairro } from './bairro';
import type { PropertyStatus } from './propertyStatus';
import type { PropertyType } from './propertyType';
import type { Zona } from './zona';

export interface PropertyInput {
  /** @minLength 1 */
  title: string;
  zona: Zona;
  neighborhood: Bairro;
  /** @minLength 1 */
  address: string;
  /** Monthly rent value in BRL */
  price: number;
  propertyType: PropertyType;
  status: PropertyStatus;
  bedrooms: number;
  bathrooms: number;
  areaM2: number;
  /** @minLength 1 */
  ownerName: string;
  /** @minLength 1 */
  ownerPhone: string;
  notes?: string;
}