# lib/api-zod/src/generated/types/propertyStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PropertyStatus = typeof PropertyStatus[keyof typeof PropertyStatus];


export const PropertyStatus = {
  captado: 'captado',
  em_negociacao: 'em_negociacao',
  disponivel: 'disponivel',
  alugado: 'alugado',
} as const;