# lib/api-zod/src/generated/types/propertyType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PropertyType = typeof PropertyType[keyof typeof PropertyType];


export const PropertyType = {
  apartamento: 'apartamento',
  casa: 'casa',
  terreno: 'terreno',
  comercial: 'comercial',
  cobertura: 'cobertura',
  rural: 'rural',
} as const;