# lib/api-zod/src/generated/types/propertyTypeCount.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyType } from './propertyType';

export interface PropertyTypeCount {
  propertyType: PropertyType;
  count: number;
}