# lib/api-zod/src/generated/types/zona.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Fixed capture zone within Porto Alegre
 */
export type Zona = typeof Zona[keyof typeof Zona];


export const Zona = {
  centro_menino_deus: 'centro_menino_deus',
  moinhos_floresta: 'moinhos_floresta',
} as const;