# lib/api-zod/src/generated/types/zonaCount.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Zona } from './zona';

export interface ZonaCount {
  zona: Zona;
  count: number;
}